Refactor, renaming
